android: Move all intent handling logic to IntentHandler to ensure that the react context exists when it is sent.
