+++ shared/react-native/android/app/src/main/java/io/keybase/ossifrage/KBPushNotifier.java
@@ -5,6 +5,7 @@ import android.app.PendingIntent;
 import android.content.Context;
 import keybase.PushNotifier;
 
+import android.os.Bundle;
 import android.support.v4.app.NotificationCompat;
 import android.support.v4.app.NotificationManagerCompat;
 
@@ -12,17 +13,26 @@ import com.dieam.reactnativepushnotification.modules.RNPushNotificationHelper;
 
 public class KBPushNotifier implements PushNotifier {
     private final Context context;
+    private Bundle bundle;
 
     public KBPushNotifier(Context ctx) {
         this.context = ctx;
     }
 
+    public void setBundle(Bundle bundle) {
+        this.bundle = bundle;
+    }
+
     public void localNotification(String ident, String msg, long badgeCount, String soundName, String convID,
             String typ) {
-        Intent open_activity_intent = new Intent(this.context, MainActivity.class);
+        Intent open_activity_intent = new Intent(context, MainActivity.class);
         open_activity_intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK | Intent.FLAG_ACTIVITY_CLEAR_TOP);
         open_activity_intent.putExtra("isNotification", true);
         open_activity_intent.putExtra("convID", convID);
+        open_activity_intent.setPackage(context.getPackageName());
+
+        this.bundle.putBoolean("userInteraction", true);
+        open_activity_intent.putExtra("notification", this.bundle);
         PendingIntent pending_intent = PendingIntent.getActivity(this.context, 0, open_activity_intent,
             PendingIntent.FLAG_UPDATE_CURRENT);
         NotificationCompat.Builder mBuilder =
